requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
environs==9.3.5
ffmpeg-python==0.2.0
future==0.18.2
idna==3.3
marshmallow==3.14.1
numpy==1.21.5
opencv-python==4.2.0.34
Pillow==8.4.0
pyTelegramBotAPI==3.7.1
pytesseract==0.3.4
python-dotenv==0.19.2
requests==2.26.0
six==1.16.0
SpeechRecognition==3.8.1
urllib3==1.26.7
